LayoutTests/accessibility/ARIA-reflection.html

<!DOCTYPE html>
<html>
<body id="body">
<script src="../resources/js-test-pre.js"></script>
<div id="content">

<div id="test" data="some data">
  <div id="firstChild"></div>
  <div id="secondChild"></div>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests ARIA IDL Attribute Reflection.");

    var element = document.getElementById("test");
    var data = element.getAttribute("data");
    var otherData = "other data";
    var currentProperty;
    var currentAttribute;

    function isElementReflectionProperty(property) {
        switch (property) {
            case "ariaActiveDescendantElement":
            case "ariaErrorMessageElement":
                return true;
        }
        return false;
    }

    function convertPropertyToAttribute(property) {
        if (isElementReflectionProperty(property))
          property = property.replace("Element", "");
        return property.replace("aria", "aria-").toLowerCase();
    }

    function testElement() {
        currentAttribute = convertPropertyToAttribute(currentProperty);

        debug("\nTest " + currentProperty + " < - > " + currentAttribute);
        shouldBeNull("element[currentProperty]");
        shouldBeNull("element.getAttribute(currentAttribute)");

        // Set the property value
        dataProperty = isElementReflectionProperty(currentProperty) ? firstChild : data;
        dataAttribute = isElementReflectionProperty(currentProperty) ? firstChild.id : data;
        element[currentProperty] = dataProperty;
        shouldBe("element.getAttribute(currentAttribute)", "dataAttribute");

        // Set the attribute value
        otherDataProperty = isElementReflectionProperty(currentProperty) ? secondChild : otherData;
        otherDataAttribute = isElementReflectionProperty(currentProperty) ? secondChild.id : otherData;
        debug("element.setAttribute(\"" + currentAttribute + "\", otherData);");
        element.setAttribute(currentAttribute, otherDataAttribute);
        shouldBe("element[currentProperty]", "otherDataProperty");
    }

    function testRole() {
        currentProperty = "role";
        testElement();
    }

    if (window.accessibilityController) {

        testRole();

        // There are 46 ARIA attributes in total.
        var count = 0;
        for (var propertyName in element) {
          if (propertyName.startsWith("aria")) {
              currentProperty = propertyName;
              testElement();
              count++;
          }
        }

        debug("\n");
        shouldBe("count", "40");

    } else {
        testFailed("Could not load accessibility controller");
    }

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
